tests/wpt/web-platform-tests/keyboard-map/keyboard-map-https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Keyboard Map basic tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

test(function() {
  assert_true(navigator.keyboard instanceof Keyboard);
}, "navigator.keyboard instanceof Keyboard");

test(function() {
  assert_equals(navigator.keyboard, navigator.keyboard);
}, "navigator.keyboard SameObject");

promise_test(function() {
  const p = navigator.keyboard.getLayoutMap();
  assert_true(p instanceof Promise);
  return p.then(function(map) {
      assert_true(map instanceof KeyboardLayoutMap);
      for (var [k,v] of map) {
        assert_true(typeof k === 'string');
        assert_true(k.length != 0);
        assert_true(typeof v === 'string');
        assert_true(v.length != 0);
      }
    });
}, "navigator.keyboard.getLayoutMap() returns a Promise<KeyboardLayoutMap> when successful");

</script>
